requirements.txt
Django==5.0.6
djangorestframework==3.15.1
pytest==8.2.2
pytestdjango==4.8.0